Fixing null reference exception for an invalid CSS rule.
If the inliner encounters an invalid CSS rule a NullReferenceException
would have occurred. The preferred behavior of this project is to just
skip the "bad" rule and continue on. This occurred when a CSS selector was
in a style sheet with no rules declared for it. An example was added to
the test suite under "InlinerShouldSkipInvalidCSSDeclarations.html".